🎨 [styles]: Add animation + Lazy loading
