+++ package.json
@@ -24,6 +24,7 @@
 		"react-redux": "^9.1.2",
 		"react-toastify": "^10.0.5",
 		"swr": "^2.2.5",
+		"tailwindcss-animated": "^1.1.2",
 		"yup": "^1.4.0"
 	},
 	"devDependencies": {
