--- src/app/(guest)/landing/_components/some_related_component.tsx
@@ -1,1 +1,0 @@
-// Delete this file after adding your own component in the directory
