+++ tailwind.config.ts
@@ -241,6 +241,6 @@ const config: Config = {
 			},
 		},
 	},
-	plugins: [],
+	plugins: [require('tailwindcss-animated')],
 };
 export default config;
